Testing done

+++ Coffeeee shop/src/main/java/com/example/coffee_shop/services/impl/ProductService.java
@@ -19,10 +19,10 @@ public class ProductService {
     public void addProduct(Product product){
         productRepository.save(product);
     }
-    public void removeProductById(long id){
+    public void removeProductById(Integer id){
         productRepository.deleteById(id);
     }
-    public Optional<Product> getProductById(long id){
+    public Optional<Product> getProductById(Integer id){
         return productRepository.findById(id);
     }
